+++ changelog.md
@@ -20,7 +20,7 @@ Parsing and sanitization improvements are always ongoing and non-breaking
 - Added `Pireps` class, module
 - Added new fields to `Station`
 - Added `from_report(report)` method to `Report`
-- Added `Aircraft`, `Turbulance`, `Icing`, `Location`, `PirepData` dataclasses in `avwx.structs`
+- Added `Aircraft`, `Turbulence`, `Icing`, `Location`, `PirepData` dataclasses in `avwx.structs`
 - Removed `priority` from `Station`
 
 ## 1.0